requirements.txt
django==3.2
black==22.6.0
isort==5.10.1
pytest-django==4.5.0